third_party/WebKit/LayoutTests/editing/selection/5354455-1.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<p>This tests whether right clicking on a paragraph break in editable content selects it.  The break should be selected on OS X, but not on Windows or Unix.  To run it manually, right click on the paragraph break after the first paragraph below.</p>
<div id="div" contenteditable="true">
<div><span id="text">The following paragraph break should be selected on OS X.</span></div>
<br>
</div>
<div id="log"></div>
<script>
function testIt(platform, expectedValue) {
    window.getSelection().removeAllRanges();
    internals.settings.setEditingBehavior(platform);

    paragraph = document.getElementById("text");
    x = paragraph.offsetParent.offsetLeft + paragraph.offsetLeft + paragraph.offsetWidth + 10;
    y = paragraph.offsetParent.offsetTop + paragraph.offsetTop + paragraph.offsetHeight / 2;
    eventSender.mouseMoveTo(x, y);
    // Give the editable region focus.
    eventSender.mouseDown();
    eventSender.mouseUp();
    // Right click on the paragraph break to select it.
    eventSender.contextClick();
    // esc key to kill the context menu.
    eventSender.keyDown("escape", null);
    assert_equals(window.getSelection().type, expectedValue);
}

if (window.eventSender && window.testRunner && window.internals) {
    test(testIt.bind(this, 'mac', 'Range'), 'mac');
    test(testIt.bind(this, 'win', 'Caret'), 'win');
    test(testIt.bind(this, 'unix', 'Caret'), 'unix');
    test(testIt.bind(this, 'android', 'Caret'), 'android');
}
</script>
